connect to mysql, add interface for servicesImpl, and add field fromWallet and toWallet in model Transaction

+++ src/main/java/com/michael/walletapi/model/dto/TransactionDTO.java
@@ -14,6 +14,6 @@ public class TransactionDTO {
     private String description;
     private double amount;
     private LocalDateTime created_at;
-    private int transaction_type_id;
-    private int wallet_id;
+    private Long transaction_type_id;
+    private Long wallet_id;
 }
